requirements.txt
streamlit>=1.30.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
